Docker support added
- Base dockerfiles added, tweaks need to be made for the prod one, as there is additional optimization that will have to be made once we have a deployment strategy

To run the express-api as a container, you will have to build, and then run the container. It links to no volumes.
TODO: Working on a dockerify vscode script

To run the database base image, run "docker-compose up -d", and "docker-compose down" respectively to run and stop it.
TODO: Clean up the setup, seeder, and migration scripts to use the docker image
TODO: Setup a vscode script to do this via the debugger